Fix (Component): Fix independent eventListeners calling wrong callback
Bug affects all version since v0.69.0.

This bug is due to preservation of variable reference when a variable is used in a loop, such  that usage of that variable in `callback` in `<element>.addEventListener('foo', callback)` now references its finalmost value before the loop exits.

The solution is to use const in the loop. By using const within the loop, a new variable is created for each loop, such that the variable is not reused in the `callback`.