Feature (Component): Add ability to use 'b-setter' attribute on component template to setup a data binding with a specified setter
As originally highlighted in #135, this feature adds the ability to declaratively define on a Component's `.template`, a data binding between an object defined among the Component's `.props`, and a custom setter defined among the Component's `.methods`.

The name of this attribute is not optimal, because it is named similarly to the original 'b-on' attribute. Hence a rename will be needed in the very near future.