Fix (environment): Initialize app with ambience specified among GET parameters
Fixes #135

We create a shadow ambience property in .props of the `menuambienceswitch` Component, which will acts as our getter/setter, thanks to #139.

The upside of this approach is that it keeps the `Memory.environment` properties unmutated. It is better than the alternative approach of creating a getter/setter on the `Memory.environment.ambiences` property directly.

The downside of this approach is that any setting of `Memory.environment.ambience` outside of this component will not update this component and the UI. However, there should be no scenario where the `Memory.environment.ambience` requires updating from outside of the context the `menuambienceswitch` component.