Fix (environment): Replace background with box-shadow for TAB-focused select element
Bug affects all versions since v0.72.0.

Bug introduced in #120.

This fixes issue where the background opacity works for :focus-visible on the select when it is unopened, but when it is opened, the background applies to the select element's option elements and is stuck to a value of 1. This issue seems to only occur on later versions of firefox.