Enhancement (environment): Redefine perfection as progressive successful characters in Student response
Closes #89

Perfection used to be defined as an exact match between the Student response and the Trainer's truth. Amendability enabled Students to make corrections to attain perfection.

Now, perfection is defined as a progressive successful match of each individual Student character to the Trainer's truth. Amends are not allowed hence the Student is constantly in the "progressive" frame of mind, i.e. "try until I get it right".

This new definition and behavior of perfection is particularly useful when paired with meditation. The philosophy of meditation is one of progress, not one of correction.

- For meditation-perfection pairing: In meditation, because the Student response is hidden from view, the Student would intuitively expect progression, not correction. Turning on perfection meets their expectation.
- Against meditation-perfection pairing: If however some students expect correction, perfection should be turned off to allow amends.

For now,  perfection will not be mandatory in meditation. But that could change in the near future.

Note that when perfection is toggled on, the training session starts over from the beginning. In perfection. Errors continue to be tracked.