Fix (meditation): Prevent margin collapsing of response component overflowing the top of Home scene in meditation
In particular the bug can be seen when the app is init, such that the Splash scene is "pushed down" relative to the Home scene, due to response component margin-top overflowing the Home scene.

This change no longer uses `display: none` on the Home scene's header component, but simply hides the header's heading text. The padding of the element pushes the response element further down the Home component, thereby preventing margin-collapsing of the response component's `margin-top` overflowing the top of the Home scene.