Enhancement (performance): Truncate Trainer Speech when book text is longer than 10000 characters
Closes #239

Previously, when book text is long e.g. > 10000 characters, the Trainer Speech requires the creation of a `<character>` element for each character for each Student response. Past 10000 characters, the app begins to show sluggishness.

Now, when book text is long e.g. > 10000 characters, the Trainer Speech is truncated and scoped only to the Speech cursor, creating only 10 Speech lines of `<character>` elements, which is much more performant. However, the Student loses the ability to review / read the entire book text during practice.

It is unfortunate that the deprecated Bubble.getPeekIndices() has to be used. There should definitely be a way to preserve performance without having to truncate the Trainer's Speech.