Fix (statistics): Fix wrong determination of a miss as a hit in perfection
Bug affects all versions since v0.84.0.

Fixes bug in #207.

Previously, a bug caused a response miss to be determined to be a hit in perfection. This incorrectly caused hit values to be higher than book text length, and the sum of values and percentages of hits, misses, amends, and other to be greater than shots value and percentage (100%). Statistics of all versions since v0.84.0 are affected by this bug.

Now, in perfection, if a response value is correct, the response is determined to be a hit, and if the value is incorrect, the value is determined to be a miss. That is, only the response cursor's value is checked.

As a beneficial side effect of this change, Trainer Speech responsiveness is enhanced, since only the response cursor's value is checked, instead of the previous implementation where all response character values are checked.