Enhancement (home): Add menu fullscreen button to Home scene to enter fullscreen programatically
Most desktop browsers include a native fullscreen function (e.g. F11 on desktop firefox, chrome, or CMD+CTRL+F on safari). However some simple users do not know how to enter fullscreen.

For the benefit of point-and-click users, a menu button is added in the Home scene for programmatic fullscreen entrance or exit.

Possible downsides:
- The UI would no longer be as straightforward as having one menu button for settings
- In programatic fullscreen, the usual ESC hotkey to navigate between Home and Environment will first exit fullscreen. Currently suppression of the ESC key by using `e.preventDefault()` doesn't seem to work in Firefox and Chrome, very likely due to the the native enforcement of ESC as the way to exit from a programmatic fullscreen.

Note: On desktop safari, in fullscreen on `document.documentElement`, `document.body`'s background applies to its actual height (leaving the html element's default white background at the bottom of the remainder of the vertical screen estate). The solution is to use `height: 100%` for both `html` and `body` elements.

This has been tested on Mac and Windows browsers, iOS and Android. Fullscreen button will show for Android browsers but not for iOS browsers.