Enhancement (speech): Show '↵' symbol when a newline (i.e. ENTER key) is expected
Closes #265

Previously, whenever a new line is expected by the Trainer, the Trainer speech shows the expectation as a 2px thin character indicator at the beginning of a line. However, this had downsides: It was difficult to notice, not intuitive for new Students, and scrolls the Trainer speech to a next line which might have been inconsistent with Student expectation that a new line (i.e. ENTER key) should be typed at the end of their current line.

This change should make it more apparant to especially new Students to type the ENTER key at the end of each line (or, in the unix understanding, the start of a new line or paragraph). Benefits: The '↵' character is more apparant, more intuitive for first time Students, and now that the Trainer speech does not scroll to the next line before a newline (i.e ENTER key) is entered would be more consistent with Student expectation that a new line should be typed at the end of the Student's current line and that the Trainer speech cursor should be mirror Student progress.