Enhancement (speech): Remove Trainer's Speech cursor fading animation
Closes #267

Previously, the Trainer's speech cursor had a fading animation. However this might distract the Student and subconsciously appear as errors, especially if the Student types with a high error rate.

Now, the animation of the Trainer cursor is removed, and it remains a static solid color. This should reduce Student subconscious distractions thereby increasing reading and typing flow.